Projection function, used in Step 3 _ Staging from the 4Smethod

This function determines the stage thresholds in the latent scale of the dimension process. Specifically, this function predicts the dimension scores corresponding to stage changes and deducts the corresponding thresholds in the dimension scale based on a correspondence between the predicted scores and the expected sum of the items. The output is a vector of the estimated stage transition thresholds.

Usage

IRT4Sstaging(D, Dss, nsim = 1000, bounds = c(-10, 10))

Arguments

D

an object of class jointLPM representing the estimated dimension model with the items

Dss

an object of class jointLPM representing the estimated proxy model between the dimension scores and the stages

nsim

number of points used in the numerical integration (Monte-Carlo). nsim should be relatively important (nsim=1000 by default).

bounds

dimension scale boundaries between which are computed a grid of the dum of the items

Value

a vector of the estimated stage transition thresholds

###@examples

References

preprint : arXiv:2407.08278

Author

Tiphaine Saulnier, Viviane Philipps and Cecile Proust-Lima